requirements.txt

altgraph==0.10.2
bdist-mpkg==0.5.0
Flask==0.10.1
flask-heroku==0.1.9
Flask-SQLAlchemy==2.1
future==0.15.2
gunicorn==19.4.5
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
macholib==1.5.1
MarkupSafe==0.23
modulegraph==0.10.4
nltk==3.1
psycopg2==2.6.1
Pillow==2.3.0
py==1.4.31
py2app==0.7.3
pyOpenSSL==0.13.1
pyparsing==2.0.1
pytest==2.8.7
python-dateutil==1.5
pytz==2013.7
requests==2.9.1
six==1.4.1
slackclient==1.0.0
stripe==1.32.0
SQLAlchemy==1.0.12
textblob==0.11.1
virtualenv==14.0.6
websocket-client==0.35.0
Werkzeug==0.11.4
xattr==0.6.4
zope.interface==4.1.1